feat: change src/router.js for most common use cases
